Move test util 'bytes' to seperate header
These utils for testing raw bytes are equally usefull when implementing
lis-types.
